Add authorization for accessing the edit or delete page by url or post request

+++ routes/campgrounds.js
@@ -34,18 +34,28 @@ router.get('/new', isLoggedIn, (req, res) => {
 
 // Provide an edit form, loading in the campground with the given ID
 router.get('/:id/edit', isLoggedIn, catchAsync(async (req, res) => {
+    const { id } = req.params;
     const campground = await Campground.findById(req.params.id);
     if(!campground) {
         req.flash('error', 'Campground was not found.');
         return res.redirect('/campgrounds');
     }    
+    if (!campground.author.equals(req.user._id)) {
+        req.flash('error', 'You do not have permission on this campground.');
+        return res.redirect(`/campgrounds/${id}`);
+    }
     res.render('campgrounds/edit', { campground });
 }));
 
 // Edit the campground with the given ID, edit route with isLoggedIn middleware
 router.put('/:id', isLoggedIn, validateCampground, catchAsync(async (req, res) => {
     const { id } = req.params;
-    const campground = await Campground.findByIdAndUpdate(id, { ...req.body.campground });
+    const campground = await Campground.findById(id)
+    if (!campground.author.equals(req.user._id)) {
+        req.flash('error', 'You do not have permission on this campground.');
+        return res.redirect(`/campgrounds/${id}`);
+    }
+    const camp = await Campground.findByIdAndUpdate(id, { ...req.body.campground });
     req.flash('success', 'Successfully updated campground.');
     res.redirect(`/campgrounds/${campground._id}`);
 }));
@@ -72,6 +82,11 @@ router.get('/:id', catchAsync(async (req, res) => {
 // Deletes campground with the given ID, protect delete route with isLoggedIn middleware
 router.delete('/:id', isLoggedIn, catchAsync(async (req, res) => {
     const { id } = req.params;
+    const campground = await Campground.findById(id)
+    if (!campground.author.equals(req.user._id)) {
+        req.flash('error', 'You do not have permission on this campground.');
+        return res.redirect(`/campgrounds/${id}`);
+    }
     await Campground.findByIdAndDelete(id);
     req.flash('success', 'Successfully deleted campground');
     res.redirect('/campgrounds');
